requirements.txt
streamlit
google-generativeai
python-dotenv

streamlit run main.py